modify:拆分举报功能为单独组件,并调整弹出举报UI展示.
